beardog_threat::threat::types::engine::ml_models

Struct MlModel

pub struct MlModel {
    pub id: String,
    pub name: String,
    pub model_type: MlModelType,
    pub accuracy: f64,
    pub last_trained: DateTime<Utc>,
    pub feature_names: Vec<String>,
}

Machine learning model structure

Represents a machine learning model used for threat detection and classification.

Fields
id: String

Model identifier

name: String

Model name

model_type: MlModelType

Model type

accuracy: f64

Model accuracy

last_trained: DateTime<Utc>

Last training timestamp

feature_names: Vec<String>

Feature names

Implementations

impl MlModel

pub fn new( id: String, name: String, model_type: MlModelType, feature_names: Vec<String>, ) -> Self

Create a new ML model

Arguments

id - Model identifier

name - Model name

model_type - Type of ML model

feature_names - Names of features used by the model

Returns

New ML model instance

Example

use beardog::threat::types::{MlModel, MlModelType};

let model = MlModel::new(
    "anomaly-detector-v1".to_string(),
    "Network Anomaly Detector".to_string(),
    MlModelType::AnomalyDetection,
    vec!["packet_count".to_string(), "byte_count".to_string()],
);

pub fn update_accuracy(&mut self, accuracy: f64)

Update model accuracy

Arguments

accuracy - New accuracy value (0.0 to 1.0)

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.update_accuracy(0.95);
assert_eq!(model.accuracy, 0.95);

pub fn mark_trained(&mut self)

Mark model as recently trained

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.mark_trained();

pub fn add_feature(&mut self, feature_name: String)

Add a feature to the model

Arguments

feature_name - Name of the feature to add

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.add_feature("request_frequency".to_string());

pub fn remove_feature(&mut self, feature_name: &str)

Remove a feature from the model

Arguments

feature_name - Name of the feature to remove

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.add_feature("test_feature".to_string());
model.remove_feature("test_feature");

pub fn feature_count(&self) -> usize

Get feature count

Returns

Number of features in the model

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.add_feature("feature1".to_string());
model.add_feature("feature2".to_string());
assert_eq!(model.feature_count(), 2);

pub fn has_feature(&self, feature_name: &str) -> bool

Check if model has a specific feature

Arguments

feature_name - Name of the feature to check

Returns

true if the model has the feature

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.add_feature("test_feature".to_string());
assert!(model.has_feature("test_feature"));

pub fn is_well_trained(&self) -> bool

Check if model is well-trained

Returns

true if model has good accuracy and features

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.update_accuracy(0.95);
model.add_feature("feature1".to_string());
assert!(model.is_well_trained());

pub fn is_recently_trained(&self) -> bool

Check if model is recently trained

Returns

true if model was trained within the last 30 days

Example

use beardog::threat::types::MlModel;

let model = MlModel::default();
assert!(model.is_recently_trained());

pub fn get_age_days(&self) -> i64

Get model age in days

Returns

Number of days since last training

Example

use beardog::threat::types::MlModel;

let model = MlModel::default();
let age = model.get_age_days();

pub fn age_days(&self) -> i64

Get age in days (alias for get_age_days)

pub fn is_high_accuracy(&self) -> bool

Check if model has high accuracy

pub fn update_training_timestamp(&mut self)

Update training timestamp

pub fn needs_retraining(&self) -> bool

Check if model needs retraining

Returns

true if model should be retrained

Example

use beardog::threat::types::MlModel;

let model = MlModel::default();
let needs_retraining = model.needs_retraining();

pub fn get_accuracy_percentage(&self) -> f64

Get accuracy percentage

Returns

Accuracy as a percentage (0-100)

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.update_accuracy(0.95);
assert_eq!(model.get_accuracy_percentage(), 95.0);

pub fn is_production_ready(&self) -> bool

Check if model is suitable for production

Returns

true if model meets production criteria

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.update_accuracy(0.95);
model.add_feature("feature1".to_string());
assert!(model.is_production_ready());

pub fn get_performance_grade(&self) -> &str

Get model performance grade

Returns

Performance grade as a string

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.update_accuracy(0.95);
assert_eq!(model.get_performance_grade(), "A");

pub fn update_model_type(&mut self, model_type: MlModelType)

Update model type

Arguments

model_type - New model type

Example

use beardog::threat::types::{MlModel, MlModelType};

let mut model = MlModel::default();
model.update_model_type(MlModelType::Classification);

pub fn clear_features(&mut self)

Clear all features

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.add_feature("feature1".to_string());
model.clear_features();
assert_eq!(model.feature_count(), 0);

pub fn set_features(&mut self, features: Vec<String>)

Set multiple features at once

Arguments

features - Vector of feature names

Example

use beardog::threat::types::MlModel;

let mut model = MlModel::default();
model.set_features(vec!["feature1".to_string(), "feature2".to_string()]);
assert_eq!(model.feature_count(), 2);